:bug: fix(paho): fixes PingHandler to timeout properly
If the local side of TCP connection is open, or `WriteTo` method does
not return error, then PingHandler can never fail because of the buggy
time check. The check seems to be a misinterpretation of the Spec which
recommends 1.5x KeepAlive for Server-side only.

This commit fixes the bug by only waiting until PingDuration time.